sirvasms/requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
configobj==4.7.2
cryptography==2.6.1
decorator==3.4.0
Django==1.11.7
django-mailbox==4.7.1
django-mathfilters==0.4.0
enum34==1.1.6
futures==3.2.0
idna==2.8
iniparse==0.4
ipaddress==1.0.16
mail-parser==3.9.2
marrow.mailer==4.0.2
marrow.util==1.2.3
meld3==0.6.10
mysql-connector==2.1.4
MySQL-python==1.2.5
numpy==1.16.2
pandas==0.24.1
perf==0.1
pycparser==2.19
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
PyJWT==1.7.1
pyliblzma==0.5.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-linux-procfs==0.4.9
pytz==2018.9
pyudev==0.15
pyxattr==0.5.1
requests==2.21.0
schedule==0.6.0
schedutils==0.4
simplejson==3.16.0
six==1.12.0
slip==0.4.0
slip.dbus==0.4.0
supervise==1.1.1
supervisor==3.1.4
twilio==6.25.0
urlgrabber==3.10
urllib3==1.24.1
xlrd==1.2.0
yum-metadata-parser==1.1.4
